backend/requirements.txt
fastmcp
langgraph
langchain-google-genai
python-dotenv
httpx
uvicorn
pydantic
